+++ lib/rtmp_stream_platform_interface.dart
@@ -27,7 +27,7 @@ abstract class RtmpStreamPlatform extends PlatformInterface {
     throw UnimplementedError('setCaptureSettings() has not been implemented.');
   }
 
-  Future<double?> create(Map<String, dynamic> params) {
+  Future<int?> create(Map<String, dynamic> params) {
     throw UnimplementedError('create() has not been implemented.');
   }
